+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 # 0.15.0-beta.3
 - [BUGFIX] Preventing the default behaviour from an event doesn't prevent the component from doing
-  the usual thing.:
+  the usual thing.
 
 # 0.15.0-beta.2
 - [BUGFIX] Fix edge case that made the component leak memory  if the component is removed after a