+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+# 0.15.0-beta.5
 - [BUGFIX] The correct behaviour when a dropdown is disabled or the tabindex is `false` should be
   to not have `tabindex` attribute, not to have a `tabindex` of -1.